Quant_Agent_Team/requirements.txt
alpaca-py>=0.43.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0
vectorbt>=0.26.0
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
